+++ src/functions/clone-deep/copy-array.ts
@@ -1,60 +1,27 @@
-import { JsonData } from '../../custom-types/json-data.ts';
-import { FunctionType, Primitive } from '../../global-types.ts';
-import { isDateArray } from '../../type-guards/is-date.ts';
-import { isFunctionArray } from '../../type-guards/is-function.ts';
-import { isObjectArray } from '../../type-guards/is-object.ts';
-import { isPrimitiveArray } from '../../type-guards/is-primitive.ts';
+import { isDate } from '../../type-guards/is-date.ts';
+import { isFunction } from '../../type-guards/is-function.ts';
+import { isObject } from '../../type-guards/is-object.ts';
+import { isPrimitive } from '../../type-guards/is-primitive.ts';
 import { copyObject } from './copy-object.ts';
 
-const copyPrimitiveOrFunctionArray = (source: Array<Primitive | FunctionType>) => {
-  let index = source.length;
-  const copy = new Array(index);
-
-  for (; index-- !== 0; ) {
-    copy[index] = source[index];
-  }
-
-  return copy;
-};
-
-const copyDateArray = (source: Array<Date>) => {
-  let index = source.length;
-  const copy = new Array(index);
-
-  for (; index-- !== 0; ) {
-    copy[index] = new Date(source[index].getTime());
-  }
-
-  return copy;
-};
-
-const copyObjectArray = (source: Array<Record<string, unknown>>) => {
-  const copy: JsonData = [];
-
-  for (let index = 0, length = source.length; index < length; index++) {
-    const newObject = copyObject(source[index]);
-    copy.push({ ...newObject });
-  }
-  console.log('copied array', copy);
-  return copy;
-};
-
 const copyArray = <T>(source: Array<T>): Array<T> => {
-  if (isPrimitiveArray(source) || isFunctionArray(source)) {
-    return copyPrimitiveOrFunctionArray(source);
-  } else if (isDateArray(source)) {
-    return copyDateArray(source);
-  } else if (isObjectArray(source)) {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    return copyObjectArray(source);
-  }
-
   let index = source.length;
   const copy = new Array(index);
 
   for (; index-- !== 0; ) {
-    copy[index] = source[index];
+    const item = source[index];
+
+    if (isPrimitive(item) || isFunction(item)) {
+      copy[index] = item;
+    } else if (isDate(item)) {
+      copy[index] = new Date(item.getTime());
+    } else if (isObject(item) && !Array.isArray(item)) {
+      copy[index] = copyObject(item);
+    } else if (Array.isArray(item)) {
+      copy[index] = copyArray(item);
+    } else {
+      copy[index] = item;
+    }
   }
 
   return copy;