+++ src/type-guards/is-object.ts
@@ -6,8 +6,4 @@ const isObject = (value: unknown): value is Record<string, unknown> => {
   return typeof value === 'object';
 };
 
-const isObjectArray = (value: unknown): value is Array<Record<string, unknown>> => {
-  return Array.isArray(value) && value.every((item) => isObject(item));
-};
-
-export { isObject, isObjectArray };
+export { isObject };