+++ src/type-guards/is-primitive.ts
@@ -5,8 +5,4 @@ const isPrimitive = (value: unknown): value is Primitive => {
   return type === 'string' || type === 'number' || type === 'boolean' || type === 'bigint';
 };
 
-const isPrimitiveArray = (value: unknown): value is Array<Primitive> => {
-  return Array.isArray(value) && value.every((item) => isPrimitive(item));
-};
-
-export { isPrimitive, isPrimitiveArray };
+export { isPrimitive };